Taking into account the move of the `decide` test suite
